+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Change Log
 
+
+## Added
+- support docxf and oform formats
+- support for the connector to work in conjunction with the plugin "Force Lowercase URLs" (https://wordpress.org/plugins/force-lowercase-urls)
+- setting authorization header
+- compatible with classic editor TinyMCE
+
 ## 1.0.2
 ## Changed
 - fixed issues for the marketplace release